% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LB4L2.R
\name{autoplot.LB4L_CD_summary}
\alias{autoplot.LB4L_CD_summary}
\title{Plot Accuracy or RT for each experimental condition of the LB4L2 dataset}
\usage{
\method{autoplot}{LB4L_CD_summary}(data, DV = "accuracy", ...)
}
\arguments{
\item{data}{An LB4L2_CD_summary data frame from the FAM package.}
}
\description{
Plot Accuracy or RT for each experimental condition of the LB4L2 dataset
}

