% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CFR_PCL.R
\name{CFR_PCL}
\alias{CFR_PCL}
\title{CFR_PCL}
\usage{
CFR_PCL(free = c(ER = 0.53, LR = 0.3, TR = 0.3, FR = 0.1, Tmin = 2, Tmax = 10,
  lambda = 0.8, alpha = 13), fixed = c(theta = 0.5, nFeat = 100, nSim = 1000,
  nList = 15, Time = 90), return_dist = TRUE)
}
\arguments{
\item{free}{}

\item{fixed}{}

\item{yoke}{}

\item{data}{}

\item{fittingRT}{}

\item{fittingAcc}{}
}
\description{
CFR_PCL
}

