% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CFR_PCR.R
\name{CFR_PCR}
\alias{CFR_PCR}
\title{CFR_PCR
An implementation of the PCR model to the FAM_CFR data set}
\usage{
CFR_PCR(free = c(ER = 0.53, LR = 0.3, Ta = 49.5, TR = 0.1, Tmin = 1, Tmax =
  30, lambda = 0.8), fixed = c(FR = 0, theta = 0.5, nFeat = 100, nSim = 1000,
  nList = 15, Time = 90), summarised = TRUE)
}
\arguments{
\item{free}{A Named vector of numeric values.}

\item{fixed}{A Named vector of numeric values.}

\item{returnDist}{Logical, controls whether a density estimate of the simulated RT distribution
 is returned along with accuracy esitmates}
}
\description{
CFR_PCR
An implementation of the PCR model to the FAM_CFR data set
}

