% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LB4L2.R, R/class_helpers.R
\name{as_LB4L}
\alias{FAM_classes}
\alias{as_LB4L}
\alias{as_LB4L_CD_summary}
\alias{as_LB4L_summary}
\alias{is.LB4L}
\alias{is.LB4L_CD_summary}
\alias{is.LB4L_summary}
\title{grep_for_error <- function(DV, variables) {
  table <- c("probability" = "acc", "RT" = "RT")
  grep(paste0('^sem_', table[DV]), variables,
       value = TRUE, perl = TRUE, ignore.case = TRUE)
}}
\usage{
as_LB4L(data, tables = attr(data, "tables"), IDvars = vapply(groups(data),
  as.character, character(1L)), experiment = attr(data, "experiment"),
  practice_tests = attr(data, "practice_tests"))

as_LB4L_summary(data, ...)

as_LB4L_CD_summary(data, ...)

is.LB4L(x)

is.LB4L_summary(x)

is.LB4L_CD_summary(x)
}
\arguments{
\item{data}{An R object that is, or can be coerced to, a data frame.}

\item{tables}{A list of named character vectors which give the name of the datasets
holding observations from each experimental phase. The datasets can take any names,
but the names of each item in the list must be one of \code{"study", "study_practice",
"test_practice"} or \code{"final"}.}

\item{IDvars}{A Character vector specifying variables that can be used to group
the data (see \code{\link{group_by}}). Defaults to the variables given by any
present grouping.}

\item{experiment}{A character vector giving the name of the experiment. Used to name
figures and any prefix any saved Rda/Rds objects.}

\item{practice_tests}{A numeric scalar giving the number of practice tests
taken.}

\item{x}{an R object}
}
\value{
The \code{as*} functions (e.g.,\code{as.LB4L_IV}, \code{as.LB4L_IV_summary})
will construct data frames (by coercing their inputs with \code{\link{as.data.frame}})
with the primary class specificed by the function (e.g., \code{as.LB4L_IV} gives the
data frame the primary class \code{"LB4L)_IV"}).

The \code{is*} functions (e.g.,\code{is.LB4L_IV}, \code{is.LB4L_IV_summary})
return logical scalars. They will return \code{TRUE} if the object has that
specific class (i.e.,\code{LB4L_IV}, \code{LB4L_IV_summary}, etc.) as it's
first value of the \code{class} attribute.
}
\description{
grep_for_error <- function(DV, variables) {
  table <- c("probability" = "acc", "RT" = "RT")
  grep(paste0('^sem_', table[DV]), variables,
       value = TRUE, perl = TRUE, ignore.case = TRUE)
}

FAM provides these helper functions to construct objects belonging to S3 classes
from the \code{\link{FAM}} package, or to quickly check if an R object
inhereits from one of the FAM package's classes.
}
\details{
It is unlikely the end-user should need to call these functions, but this page
serves to document their behavior should they need to be used externally from
the \code{\link{FAM}} package, and to give more information on the structure
of the S3 classes implemented by the \code{\link{FAM}} package.
}
\section{Functions}{
\itemize{
\item \code{as_LB4L}: Creates a data frame with the primary class "LB4L"

\item \code{as_LB4L_summary}: Creates a data frame with the primary class "LB4L_summary"

\item \code{as_LB4L_CD_summary}: Creates a data frame with the primary class "LB4L_CD_summary"

\item \code{is.LB4L}: Checks if an R object has \code{LB4L} as it's
first value of the \code{class} attribute.

\item \code{is.LB4L_summary}: Checks if an R object has \code{LB4L_summary} as it's
first value of the \code{class} attribute.
#' @export

\item \code{is.LB4L_CD_summary}: Checks if an R object has \code{LB4L_CD_summary} as it's
first value of the \code{class} attribute.
#' @export
}}

