% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_stays.R
\name{family_stays}
\alias{family_stays}
\title{Title Manipulation on the Hotel Reservations}
\usage{
family_stays(rsvn, pos_col1, pos_col2, night1, night2, target)
}
\arguments{
\item{rsvn}{A dataset}

\item{pos_col1}{Column with only positive values keepped}

\item{pos_col2}{Column with only positive values keepped}

\item{night1}{Column recording the number of nights reserved}

\item{night2}{Column recording the number of nights reserved}

\item{target}{Predicted variables}
}
\value{
Data frame with 2 columns only
}
\description{
Title Manipulation on the Hotel Reservations
}
\examples{
rsvn = tibble::tibble(no_of_adults = c(-1,-2,5,8,10), no_of_children = c(-2,0,5,6,7),
 no_of_weekend_nights = c(1,0,2,7,1), no_of_week_nights = c(0,1,2,3,6),
 booking_status = c('d','e','e','d','e'))
family_stays(rsvn, no_of_adults, no_of_children, no_of_weekend_nights,
no_of_week_nights, booking_status)
}
