% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LB4L_PCL.R
\name{fitLB4L}
\alias{fitLB4L}
\title{Title}
\usage{
fitLB4L(model, inpar = FALSE, debugLevel = 0)
}
\arguments{
\item{model}{}

\item{inpar}{}

\item{debugLevel}{}
}
\value{
A Model list, updated with a list of  the fitting results and model predictions
}
\description{
Title
}
\examples{
model <- list(par = c(ER=.52,LR=.2,TR =.03, F1=.05,space=.03),
               fn = PCLss, method="Nelder-Mead", itnmax = 1000,
               fix= c(theta=.5,nFeat=100,nSim=1000,nList=15),
               IVdata= data, jointData = data, condData = data,
               fitting = TRUE, name = 'std_ss'))
 results <- fitLB4L(model,inpar = TRUE)
}

