% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LB4L2_PCR.R
\name{fitPCR}
\alias{fitPCR}
\title{fitPCR}
\usage{
fitPCR(parameters = NULL, PCRparams_obj, objective_fcn, error_fcn, ...)
}
\arguments{
\item{objective_fcn}{E.g., LB4L2_PCR}

\item{...}{}
}
\value{

}
\description{
fitPCR
}

