% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SCKT_summaries.R
\name{IVsummary}
\alias{IVsummary}
\title{IVsummary}
\usage{
IVsummary(data, grouping.vars = c("subject", "group", "practice",
  "other_type"), measure.vars = c("prac_score", "final_score"), fn = "mean")
}
\description{
IVsummary
}
\examples{
IVdata_grouped <- IVsummary(SCKT_allSs,
                            grouping.vars =c("subject","group",
                             "half","practice","other_type"))
}

