% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LB4L2_PCR.R
\name{LB4L2_PCR}
\alias{LB4L2_PCR}
\title{LBL42 PCR}
\usage{
LB4L2_PCR(PCR_model_obj)
}
\arguments{
\item{PCR_model_obj}{}
}
\value{
A data frame
}
\description{
Implementation of the PCR model of LB4L2 experiment
}
\examples{
model_params <- initPCRparams(params = list(ER=.6, LR=.2, FR=.1, TR=.15, TV = .05),
                   distribution = "beta", nItems = 20, nSims = 1000,
                   nFeatures = 100, time = 10)
results <- LB4L2_PCR(model_params)

}

