% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LB4L2_PCR.R
\name{LB4L2_PCR_erf}
\alias{LB4L2_PCR_erf}
\title{LB4L2_PCR Error Function}
\usage{
LB4L2_PCR_erf(results, IV_obs, joint_obs, likelihood = c("RT", "accuracy",
  "all"), ...)
}
\arguments{
\item{likelihood}{}
}
\value{

}
\description{
LB4L2_PCR Error Function
}

