% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LB4L_PCR.R
\name{LB4L_PCRss}
\alias{LB4L_PCRss}
\title{Title}
\usage{
LB4L_PCRss(free = c(ER = 0.52, LR = 0.2, TR = 0.03, F1 = 0.05, space = 0.03),
  fixed = c(theta = 0.5, nFeat = 100, nSim = 1000, nList = 15, Tmin = NA, Tmax
  = NA, lambda = NA, Time = NA), IVdata = filter(LB4L_IV(LB4L_allSs)$subject,
  subject == 17), jointData = filter(LB4L_joint(LB4L_allSs)$subject, subject
  == 17), fitting = FALSE, ...)
}
\arguments{
\item{free}{Named vector of free parameters. See Details for possible names.}

\item{fixed}{Named vector of free parameters. See Details for possible names.}

\item{IVdata}{}

\item{fitting}{Controls  the return value.  See Value for details.}

\item{jointdata}{}
}
\value{
If \code{fitting = TRUE}, scalar vector is returned,
 useful for optimization routine. If \code{fittting = FALSE}, list of data frames with predicted joint,
 conditional, and IV condition-level predictions is returned.
}
\description{
Title
}
\examples{
LB4L_PCRss(free= c(ER=.53,LR=.15,TR =.1, F1=.1,space=.03),
           fixed = c(theta=.5,nFeat=100,nSim=1000,nList=15,
                    Tmin=NA, Tmax=NA, lambda=NA,Time=NA),
                    fitting=FALSE)
}

