% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CFR_PCR.R
\name{RTdist}
\alias{RTdist}
\title{RTdist: Normalized Density Estimates}
\usage{
RTdist(RT, time = 90)
}
\arguments{
\item{RT}{}

\item{time}{Totall test duration measured in seconds. Density is estimated
 at every tenth of a second between .1 and \code{time}}
}
\value{
A data frame
}
\description{
Estimated Gaussian kernel density at each 1/10 second of recall test period,
using a bandwidth of 1. Density results are normalized so that the sum
of the densities at each output positions is 1.
}
\examples{
preds <- CFR_PCR(summarised=FALSE)
 dist <- preds \%>\% group_by(class,obsOrder) \%>\% RTdist()
}

