% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LB4L2.R
\name{summary.LB4L}
\alias{summary.LB4L}
\title{Summarizing the raw LB4L2 dataset by the experimentally manipulated factors}
\usage{
\method{summary}{LB4L}(data, DV = "accuracy", level = "subject",
  given_practice = FALSE, given_final = (DV == "RT"))
}
\arguments{
\item{data}{LB4L or LB4L2 data set from the FAM package.}

\item{DV}{A character vector of length one, used for determing the which dependent
variable should be summarised in each experimental condition. If the value is
\code{"accuracy"} (the default) then the percent of items correctly recalled
is returned for each condition. If the value is \code{"RT"}, then the first-press
latency in each experimental condition is returned, conditioned on accuracy of
the response.}

\item{level}{A character vector of length one, used for determing the level
of the summary. If the value is \code{"subject"} (the default) then the means
for each subject are returned, and if the value is \code{"group"} then the means
of each experimental condition is returned, aggregated over individual subjects.}

\item{given_practice}{A numeric vector listing which practice tests should be
used as grouping variables for calculating conditional final test averages. A logical
vector of length one (i.e., \code{TRUE} or \code{FALSE} can be used to indicate
all practice tests should be used, or no practice test should be used.}

\item{given_final}{A logical vector of length one, indicating whether or not
group observations for specific targets according to the accuracy outcome on the final test.
Only used when the value of the \code{DV} argument is \code{"RT"}.}
}
\description{
The LB4L2 experiment data is stored in 4 different data sets, each corresponding
to a separate phase of the LB4L2 experiment (see LB4L2 datasets for more info).
This is a method for the S3 generic function \code{\link{summary}} which reports
either the subject-level or condition-levels means for each experimental condition.
}
\note{
Trials where no keys were pressed by the participant are excluded from
the RT summary.
}

