% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LB4L2_PCR.R
\name{LB4L2_PCR}
\alias{LB4L2_PCR}
\title{LBL42 PCR}
\usage{
LB4L2_PCR(PCR_model_obj, groups = 1)
}
\arguments{
\item{PCR_model_obj}{A PCRparams object, as generated by \code{initPCR}}

\item{groups}{A numeric scalar indicating how many independtn groups to
make predictions for}
}
\value{
A list of PCR objects, with as many elements as there are \code{groups}.
Each list element holds the PCR objects created for each condition.
}
\description{
Implementation of the PCR model of LB4L2 experiment
}
\examples{
model_params <- initPCR(params = list(ER=.6, LR=.2, FR=.1, TR=.15, TV = .05),
                   distribution = "beta", nItems = 20, nSims = 1000,
                   nFeatures = 100, time = 10)
results <- LB4L2_PCR(model_params)

}

