% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LB4L2.R
\name{summary.LB4L}
\alias{summary.LB4L}
\title{Summarizing the raw LB4L2 dataset by the experimentally manipulated factors}
\usage{
\method{summary}{LB4L}(data, level = "subject")
}
\arguments{
\item{data}{An LB4L data set from the FAM package.}

\item{level}{A character vector of length one, used for determing the level
of the summary. If the value is \code{"subject"} (the default) then the summary statistics
for each subject are returned. If the value is \code{"group"} then the means
of each experimental condition is returned, pooled across individual subjects.}
}
\description{
The LB4L2 experiment data is stored in 4 different data sets, each corresponding
to a separate phase of the LB4L2 experiment (see LB4L2 datasets for more info).
This is a method for the S3 generic function \code{\link{summary}} which reports
either the subject-level or condition-levels means for each experimental condition.
}
\note{
Trials where no keys were pressed by the participant are excluded from
the RT summary.
}

